LayoutTests/storage/indexeddb/create-and-remove-object-store-expected.txt

Test IndexedDB's create and removeObjectStore

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'webkitIndexedDB' in window is true
PASS webkitIndexedDB == null is false
webkitIndexedDB.open('name')
PASS 'onsuccess' in result is true
PASS 'onerror' in result is true
PASS 'readyState' in result is true
An event should fire shortly...

openSuccess():
Success event fired:
PASS 'result' in event is true
PASS 'code' in event is false
PASS 'message' in event is false
PASS 'source' in event is true
PASS event.source != null is true
PASS 'onsuccess' in event.target is true
PASS 'onerror' in event.target is true
PASS 'readyState' in event.target is true
PASS event.target.readyState is event.target.DONE

db = event.result
Trying create
PASS code is webkitIDBDatabaseException.NOT_ALLOWED_ERR
Trying remove
PASS code is webkitIDBDatabaseException.NOT_ALLOWED_ERR
result = db.setVersion('version 1')
Trying create
PASS code is webkitIDBDatabaseException.NOT_ALLOWED_ERR
Trying remove
PASS code is webkitIDBDatabaseException.NOT_ALLOWED_ERR
Success event fired:
PASS 'result' in event is true
PASS 'code' in event is false
PASS 'message' in event is false
PASS 'source' in event is true
PASS event.source != null is true
PASS 'onsuccess' in event.target is true
PASS 'onerror' in event.target is true
PASS 'readyState' in event.target is true
PASS event.target.readyState is event.target.DONE

Deleted all object stores.
db.createObjectStore('tmp')
Adding 'tmp' again
PASS code is webkitIDBDatabaseException.CONSTRAINT_ERR
trans = db.transaction({mode: webkitIDBTransaction.READ_WRITE})
trans.objectStore('tmp').get(0)
Trying create
PASS code is webkitIDBDatabaseException.NOT_ALLOWED_ERR
Trying remove
PASS code is webkitIDBDatabaseException.NOT_ALLOWED_ERR
PASS successfullyParsed is true

TEST COMPLETE
